ai/test: extractFunctions関数のテストケースを追加し、テストファイルをanalyzer_test.goに統合
